+++ App_Data/jobs/triggered/updatePerformanceAvailabilities/run.ts
@@ -35,7 +35,7 @@ ttts.service.itemAvailability.updatePerformanceAvailabilities({
     startThrough: startThrough,
     ttl: parseInt(PERFORMANCE_AVAILABILITY_EXPIRES_IN_SECONDS_STR, 10)
 })(
-    new ttts.repository.Stock(ttts.mongoose.connection),
+    new ttts.repository.Stock(redisClient),
     new ttts.repository.Performance(ttts.mongoose.connection),
     new ttts.repository.itemAvailability.Performance(redisClient)
 )