+++ CHANGELOG.md
@@ -12,6 +12,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 - パフォーマンスコレクションからのリレーション削除対応に合わせて、自動生成パフォーマンスの型を調整
 - jsdoc削除
+- 予約リポジトリのmongooseモデルをprivate化
+- パフォーマンスリポジトリのmongooseモデルをprivate化
+- 新しいRedis在庫リポジトリで再構築
 
 ### Deprecated
 
@@ -20,6 +23,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - 劇場、スクリーン、作品、券種、券種グループコレクションを削除(一時的に動的データ管理を不可能に変更)
 - GMO通知コレクションを削除
 - SendGridイベントコレクションを削除
+- 在庫作成ジョブを削除
 
 ### Fixed